[refactor] [tests] Refactor CliFrontend mocking into utility class
